Create PR Build
Using GitHub actions, create an action that runs when a new pr is submitted or new commits are added. This is useful for when someone is building and it errors on their machine only, they will be able to test and other people will be able to receive the jar that came from the changes.